dev(release): use ts-node, add top-level shortuct
