+++ dev/release/README.md
@@ -7,13 +7,8 @@ on our release process and how this tool is used.
 To see all available steps:
 
 ```sh
-yarn build
 yarn run release help # add 'all' to see test commands as well
 ```
 
-Run `yarn run build` to build the tool (_required_ alongside `yarn run release` to make
-sure you are using the latest version of the tool), and `yarn run watch` to build on any
-changes to files.
-
-Before using this tool, please also verify that the [release configuration](./release-config.jsonc)
+Before using this tool, please verify that the [release configuration](./release-config.jsonc)
 is set up correctly.
