+++ package.json
@@ -30,7 +30,8 @@
     "storybook": "start-storybook -p 9001 -c .storybook -s ui/assets",
     "build-storybook": "build-storybook -c .storybook -s ui/assets",
     "cover-storybook": "nyc --hook-require=false yarn jest .storybook/coverage",
-    "deduplicate": "yarn-deduplicate -s fewer"
+    "deduplicate": "yarn-deduplicate -s fewer",
+    "release": "cd dev/release && yarn run release"
   },
   "browserslist": [
     "last 1 version",
